refactor: reorganize providers into categorized subdirectories
Move provider implementations from /services to /providers with clear
subdirectories:
- /providers/players/ - LocalPlayer, RemotePlayer, jsonRpcWebSocket
- /providers/lyrics/ - LrclibLyricsProvider, LocalLyricsProvider
- /providers/artwork/ - ITunesArtworkProvider, UnsplashArtworkProvider

This improves code organization and distinguishes provider
implementations from application services in /core/services.

Updated all imports in:
- config/providers.ts (7 import statements)
- tests/unit/localPlayer.test.ts
- tests/unit/lrclibLyricsProvider.test.ts
- tests/unit/settings.test.ts

All 155 tests passing.
