+++ client/src/config/providers.ts
@@ -12,7 +12,7 @@ export const BUILTIN_PROVIDER_CONFIGS = {
       name: "Local",
       description: "Local player",
       load: async (): Promise<Player> => {
-        const { LocalPlayer } = await import("@/services/localPlayer");
+        const { LocalPlayer } = await import("@/providers/players/localPlayer");
         return LocalPlayer.getInstance();
       },
     },
@@ -21,7 +21,9 @@ export const BUILTIN_PROVIDER_CONFIGS = {
       name: "Server",
       description: "Remote player",
       load: async (): Promise<Player> => {
-        const { RemotePlayer } = await import("@/services/remotePlayer");
+        const { RemotePlayer } = await import(
+          "@/providers/players/remotePlayer"
+        );
         return RemotePlayer.getInstance();
       },
     },
@@ -34,7 +36,7 @@ export const BUILTIN_PROVIDER_CONFIGS = {
         "Community-driven lyrics database with synchronized lyrics support",
       load: async (): Promise<LyricsProvider> => {
         const { LrclibLyricsProvider } = await import(
-          "@/services/lrclibLyricsProvider"
+          "@/providers/lyrics/lrclibLyricsProvider"
         );
         return new LrclibLyricsProvider();
       },
@@ -47,7 +49,7 @@ export const BUILTIN_PROVIDER_CONFIGS = {
       description: "Album artwork from iTunes Search API",
       load: async (): Promise<ArtworkProvider> => {
         const { ITunesArtworkProvider } = await import(
-          "@/services/itunesArtworkProvider"
+          "@/providers/artwork/itunesArtworkProvider"
         );
         return new ITunesArtworkProvider();
       },
@@ -58,7 +60,7 @@ export const BUILTIN_PROVIDER_CONFIGS = {
       description: "High-quality random images optimized for your screen size",
       load: async (): Promise<ArtworkProvider> => {
         const { UnsplashArtworkProvider } = await import(
-          "@/services/unsplashArtworkProvider"
+          "@/providers/artwork/unsplashArtworkProvider"
         );
         return new UnsplashArtworkProvider();
       },
