+++ client/src/providers/artwork/itunesArtworkProvider.ts
